feat: change to GameView
